Introducing Money object to encapsulate amounts. This is to fix the problem of the api returning whole number but the app expecting, wrongly, decimals. Introduced reusable MoneyView widget to encapsulate formatting of Money objects. Also fixing some unit tests. there is one failing acceptance test displaying error for getting balance. will fix next

+++ app/src/androidTest/java/com/savvasdalkitsis/mondo/test/data/TestResponses.java
@@ -2,11 +2,11 @@ package com.savvasdalkitsis.mondo.test.data;
 
 public class TestResponses {
 
-    public static String balance(double balance, TestCurrency currency) {
+    public static String balance(int balance, TestCurrency currency) {
         return "{\"balance\": " + balance + ", \"currency\": \"" + currency.apiName() + "\"}";
     }
 
-    public static String spentToday(double amount, TestCurrency currency) {
+    public static String spentToday(int amount, TestCurrency currency) {
         return "{\"spent_today\": " + amount + ", \"currency\": \"" + currency.apiName() + "\"}";
     }
 }
